+++ apps/meteor/ee/server/apps/communication/endpoints/appsCountHandler.ts
@@ -1,9 +1,9 @@
 import type { AppManager } from '@rocket.chat/apps-engine/server/AppManager';
+import { License } from '@rocket.chat/license';
 
 import { API } from '../../../../../app/api/server';
 import type { SuccessResult } from '../../../../../app/api/server/definition';
 import { getInstallationSourceFromAppStorageItem } from '../../../../../lib/apps/getInstallationSourceFromAppStorageItem';
-import { getAppsConfig } from '../../../../app/license/server/license';
 import type { AppsRestApi } from '../rest';
 
 type AppsCountResult = {
@@ -23,7 +23,7 @@ export const appsCountHandler = (apiManager: AppsRestApi) =>
 				const manager = apiManager._manager as AppManager;
 
 				const apps = manager.get({ enabled: true });
-				const { maxMarketplaceApps, maxPrivateApps } = getAppsConfig();
+				const { maxMarketplaceApps, maxPrivateApps } = License.getAppsConfig();
 
 				return API.v1.success({
 					totalMarketplaceEnabled: apps.filter((app) => getInstallationSourceFromAppStorageItem(app.getStorageItem()) === 'marketplace')