+++ apps/meteor/server/services/room/service.ts
@@ -23,15 +23,13 @@ export class RoomService extends ServiceClassInternal implements IRoomService {
 			throw new Error('no-permission');
 		}
 
-		const user = await Users.findOneById<Pick<IUser, 'username'>>(uid, {
-			projection: { username: 1 },
-		});
+		const user = await Users.findOneById(uid);
 		if (!user?.username) {
 			throw new Error('User not found');
 		}
 
 		// TODO convert `createRoom` function to "raw" and move to here
-		return createRoom(type, name, user.username, members, false, readOnly, extraData, options) as unknown as IRoom;
+		return createRoom(type, name, user, members, false, readOnly, extraData, options) as unknown as IRoom;
 	}
 
 	async createDirectMessage({ to, from }: { to: string; from: string }): Promise<{ rid: string }> {