+++ packages/core-typings/src/index.ts
@@ -39,7 +39,6 @@ export * from './IUserSession';
 export * from './IUserStatus';
 export * from './IUser';
 
-export * from './ee/ILicense/ILicense';
 export * from './ee/IAuditLog';
 
 export * from './import';
@@ -135,3 +134,5 @@ export * from './ICustomOAuthConfig';
 
 export * from './IModerationReport';
 export * from './CustomFieldMetadata';
+
+export * as Cloud from './cloud';